rename and move from bootstrap_check to agent directory
