+++ logstash-core/lib/logstash/agent.rb
@@ -25,6 +25,7 @@ require "logstash/config/pipeline_config"
 require "logstash/pipeline_action"
 require "logstash/state_resolver"
 require "logstash/pipelines_registry"
+require "logstash/persisted_queue_config_validator"
 require "stud/trap"
 require "uri"
 require "socket"
@@ -94,7 +95,7 @@ class LogStash::Agent
 
     initialize_geoip_database_metrics(metric)
     
-    @bootstrap_check_pq = LogStash::BootstrapCheck::PersistedQueueConfig.new
+    @pq_config_validator = LogStash::PersistedQueueConfigValidator.new
 
     @dispatcher = LogStash::EventDispatcher.new(self)
     LogStash::PLUGIN_REGISTRY.hooks.register_emitter(self.class, dispatcher)
@@ -185,7 +186,7 @@ class LogStash::Agent
       end
     end
 
-    @bootstrap_check_pq.check(@pipelines_registry.running_user_defined_pipelines, results.response)
+    @pq_config_validator.check(@pipelines_registry.running_user_defined_pipelines, results.response)
 
     converge_result = resolve_actions_and_converge_state(results.response)
     update_metrics(converge_result)
